PsqlSaleDeNorm has an error with id not incrementing correctly (one inc per total sale), cna make it work but tkaes long time

+++ CPSAssignment2.Benchmark/Models/MasterCustomer.cs
@@ -17,7 +17,7 @@ namespace CPSAssignment2.Benchmark.Models
         public long Account2 { get; set; }
         public long Account3 { get; set; }
 
-        public static List<MasterCustomer> GenerateCustomers(int amount, string optionalDomain = null, int optionalIdstart = 0)
+        public static List<MasterCustomer> GenerateCustomers(int amount, string optionalDomain = null, long optionalIdstart = 0)
         {
             List<MasterCustomer> Customers = new List<MasterCustomer>();
             string optdom = "-gen";
